backend/requirements.txt
fastapi
uvicorn
pymysql
python-dotenv
passlib[bcrypt]
pyjwt